requirements.txt
# Core AI / LLM
openai>=1.0.0,<2.1.0
langchain-core>=0.3.53,<1.0.0
langchain-community>=0.3.21,<1.0.0
langchain-openai>=0.3.14,<1.0.0
langchain>=0.3.23,<1.0.0
faiss-cpu>=1.7.4

# Web scraping & news
requests>=2.31.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
lxml>=4.9.0

# Data & finance
pandas>=2.0.0
yfinance>=0.2.36
pytrends>=4.9.2

# NLP
spacy>=3.7.0
textblob>=0.17.1

# Reddit
praw>=7.7.0

# Streamlit & PDF
streamlit>=1.32.0
matplotlib>=3.8.0
xhtml2pdf>=0.2.11

# Env
python-dotenv>=1.0.0
